rename `kernel` module to `routines`
