+++ honeycomb-kernels/src/grisubal/mod.rs
@@ -52,20 +52,20 @@
 
 // ------ MODULE DECLARATIONS
 
-pub(crate) mod kernel;
 pub(crate) mod model;
+pub(crate) mod routines;
 pub(crate) mod timers;
 
 // ------ IMPORTS
 
 use crate::grisubal::{
-    kernel::{
+    model::{Boundary, Geometry2},
+    routines::{
         clip_left, clip_right, compute_intersection_ids, compute_overlapping_grid,
         detect_orientation_issue, generate_edge_data, generate_intersection_data,
         group_intersections_per_edge, insert_edges_in_map, insert_intersections,
         remove_redundant_poi,
     },
-    model::{Boundary, Geometry2},
     timers::{finish, start_timer, unsafe_time_section},
 };
 use honeycomb_core::{
